pfm/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.20.0
telethon>=1.36.0
actualpy>=0.7.0
python-dotenv>=1.0.0
